refactor/simplify concretization
