+++ src/audit/overprovisioned_secrets.rs
@@ -27,7 +27,7 @@ impl Audit for OverprovisionedSecrets {
                 continue;
             };
 
-            todo!()
+            // todo!()
         }
 
         Ok(vec![])
