+++ src/finding/mod.rs
@@ -4,7 +4,6 @@ use std::{borrow::Cow, sync::LazyLock};
 
 use anyhow::{anyhow, Result};
 use clap::ValueEnum;
-use locate::Locator;
 use regex::Regex;
 use serde::Serialize;
 use terminal_link::Link;
@@ -13,8 +12,6 @@ use crate::{
     models::{CompositeStep, JobExt, Step},
     registry::InputKey,
 };
-
-pub(crate) mod locate;
 
 /// Represents the expected "persona" that would be interested in a given
 /// finding. This is used to model the sensitivity of different use-cases
@@ -192,11 +189,38 @@ impl<'w> SymbolicLocation<'w> {
         self,
         document: &'w impl AsRef<yamlpath::Document>,
     ) -> Result<Location<'w>> {
-        let feature = Locator::new().concretize(document, &self)?;
+        let document = document.as_ref();
+
+        // If we don't have a path into the workflow, all
+        // we have is the workflow itself.
+        let feature = if self.route.components.is_empty() {
+            document.root()
+        } else {
+            let mut builder = yamlpath::QueryBuilder::new();
+
+            for component in &self.route.components {
+                builder = match component {
+                    RouteComponent::Key(key) => builder.key(key.clone()),
+                    RouteComponent::Index(idx) => builder.index(*idx),
+                }
+            }
+
+            let query = builder.build();
+
+            document.query(&query)?
+        };
 
         Ok(Location {
             symbolic: self,
-            concrete: feature,
+            concrete: Feature {
+                location: ConcreteLocation::from(&feature.location),
+                feature: document.extract_with_leading_whitespace(&feature),
+                comments: document
+                    .feature_comments(&feature)
+                    .into_iter()
+                    .map(Comment)
+                    .collect(),
+            },
         })
     }
 }
